Move mainboard-specific setup all into mainboard.rs, and move common setup to soc/amd/common/boot.
