+++ src/soc/amd/common/boot/src/acpi.rs
